ADD: Implement CricketMatch and CricketMatchDetail classes, update Team ID types, and enhance MatchController for match creation

+++ crick-dao/src/main/java/com/shouryashrey/crick_dao/repos/TeamRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Integer> {
+public interface TeamRepository extends JpaRepository<Team, Long> {
     Optional<Team> findByTeamName(String teamName);
     boolean existsByTeamName(String teamName);
     boolean existsByShortName(String shortName);
